Fixed bootstrap styling for skills  section
Styling for skills page so bootstrap orders columns properly on resize, and made it look cleaner. Changed profile picture to look smaller, and added a border so I can have icons that link to external pages(github, LinkedIn, and email). Seperated that from the footer so I can use it in the picture's border in about me