Reorganized Components
Came across a bug in nav bar that would set the state value in navbar many times since it was a scroll event handler. To fix this, I am going to use refs in my applicatoin and remove the npm package I am using for scroll url changes. This required a new component to encomposs everything, not having about me be the parent anf eveyrthing else being the child.